+++ frontend/appflowy_tauri/src-tauri/Cargo.toml
@@ -120,14 +120,14 @@ custom-protocol = ["tauri/custom-protocol"]
 # To switch to the local path, run:
 # scripts/tool/update_collab_source.sh
 # ⚠️⚠️⚠️️
-collab = { version = "0.2", git = "https://github.com/AppFlowy-IO/AppFlowy-Collab", rev = "925c804e2eb4a01f1308e08136a8e14e9e2baf2e" }
-collab-entity = { version = "0.2", git = "https://github.com/AppFlowy-IO/AppFlowy-Collab", rev = "925c804e2eb4a01f1308e08136a8e14e9e2baf2e" }
-collab-folder = { version = "0.2", git = "https://github.com/AppFlowy-IO/AppFlowy-Collab", rev = "925c804e2eb4a01f1308e08136a8e14e9e2baf2e" }
-collab-document = { version = "0.2", git = "https://github.com/AppFlowy-IO/AppFlowy-Collab", rev = "925c804e2eb4a01f1308e08136a8e14e9e2baf2e" }
-collab-database = { version = "0.2", git = "https://github.com/AppFlowy-IO/AppFlowy-Collab", rev = "925c804e2eb4a01f1308e08136a8e14e9e2baf2e" }
-collab-plugins = { version = "0.2", git = "https://github.com/AppFlowy-IO/AppFlowy-Collab", rev = "925c804e2eb4a01f1308e08136a8e14e9e2baf2e" }
-collab-user = { version = "0.2", git = "https://github.com/AppFlowy-IO/AppFlowy-Collab", rev = "925c804e2eb4a01f1308e08136a8e14e9e2baf2e" }
-collab-importer = { version = "0.1", git = "https://github.com/AppFlowy-IO/AppFlowy-Collab", rev = "925c804e2eb4a01f1308e08136a8e14e9e2baf2e" }
+collab = { version = "0.2", git = "https://github.com/AppFlowy-IO/AppFlowy-Collab", rev = "287cf7f480ef2f31364eaf33aab81e922fdb43df" }
+collab-entity = { version = "0.2", git = "https://github.com/AppFlowy-IO/AppFlowy-Collab", rev = "287cf7f480ef2f31364eaf33aab81e922fdb43df" }
+collab-folder = { version = "0.2", git = "https://github.com/AppFlowy-IO/AppFlowy-Collab", rev = "287cf7f480ef2f31364eaf33aab81e922fdb43df" }
+collab-document = { version = "0.2", git = "https://github.com/AppFlowy-IO/AppFlowy-Collab", rev = "287cf7f480ef2f31364eaf33aab81e922fdb43df" }
+collab-database = { version = "0.2", git = "https://github.com/AppFlowy-IO/AppFlowy-Collab", rev = "287cf7f480ef2f31364eaf33aab81e922fdb43df" }
+collab-plugins = { version = "0.2", git = "https://github.com/AppFlowy-IO/AppFlowy-Collab", rev = "287cf7f480ef2f31364eaf33aab81e922fdb43df" }
+collab-user = { version = "0.2", git = "https://github.com/AppFlowy-IO/AppFlowy-Collab", rev = "287cf7f480ef2f31364eaf33aab81e922fdb43df" }
+collab-importer = { version = "0.1", git = "https://github.com/AppFlowy-IO/AppFlowy-Collab", rev = "287cf7f480ef2f31364eaf33aab81e922fdb43df" }
 
 
 # Working directory: frontend
